change torch tensor type

+++ data/cancer_data.py
@@ -55,6 +55,6 @@ class cancer_data(classification_data):
         y_val = y_val.to(self.args.device)
         data = {'x_train':x_train, "y_train":y_train, "x_val":x_val, "y_val":y_val}
 
-        cancer_train= TensorDataset(torch.FloatTensor(data['x_train']),torch.FloatTensor(data['y_train']))
-        cancer_test= TensorDataset(torch.FloatTensor(data['x_val']),torch.FloatTensor(data['y_val']))
+        cancer_train= TensorDataset(data['x_train'],data['y_train'])
+        cancer_test= TensorDataset(data['x_val'],data['y_val'])
         return (cancer_train,cancer_test)
