+++ data/iris_data.py
@@ -56,6 +56,6 @@ class iris_data(classification_data):
         y_val = y_val.to(self.args.device)
         data = {'x_train':x_train, "y_train":y_train, "x_val":x_val, "y_val":y_val}
 
-        iris_train= TensorDataset(torch.FloatTensor(data['x_train']),torch.LongTensor(data['y_train']))
-        iris_test= TensorDataset(torch.FloatTensor(data['x_val']),torch.LongTensor(data['y_val']))
+        iris_train= TensorDataset(data['x_train'],data['y_train'])
+        iris_test= TensorDataset(data['x_val'],data['y_val'])
         return (iris_train,iris_test)
